# packages/api-spec/tsp-output/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: DDD Clean Twitter API
  version: 0.0.0
...
paths:
  /api/auth/login:
    post:
      operationId: Auth_login
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Api.LoginResponse'
                  - $ref: '#/components/schemas/Api.ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Api.LoginRequest'
  /api/auth/me:
    get:
      operationId: Auth_me
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Api.MeResponse'
                  - $ref: '#/components/schemas/Api.ErrorResponse'
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    Api.ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    Api.LoginRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
    Api.LoginResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/Api.User'
    Api.MeResponse:
      type: object
      required:
        - sub
        - email
      properties:
        sub:
          type: string
        email:
          type: string
...
    Api.User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:3000
    description: Local server
    variables: {}